backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.124.4",
    "passlib>=1.7.4",
    "python-jose[cryptography]>=3.5.0",
    "sqlalchemy>=2.0.45",
    "uvicorn>=0.38.0",
]
